change(it-hygiene): move ports to networks directory and move processes sub tab directory to main directory of processes
