+++ plugins/main/public/components/overview/it-hygiene/networks/inventories/index.ts
@@ -1,3 +1,4 @@
 export * from './interfaces/inventory';
 export * from './networks/inventory';
 export * from './protocols/inventory';
+export * from './ports/inventory';
