+++ plugins/main/public/components/overview/it-hygiene/networks/inventories/ports/managed-filters.ts
