+++ plugins/main/public/components/overview/it-hygiene/networks/inventories/ports/table-columns.ts
