--- plugins/main/public/components/overview/it-hygiene/processes/inventories/index.ts
@@ -1,2 +1,0 @@
-export * from './processes/inventory';
-export * from './ports/inventory';
