--- plugins/main/public/components/overview/it-hygiene/processes/inventories/processes/inventory.tsx
@@ -1,27 +1,0 @@
-import React from 'react';
-import {
-  SystemInventoryStatesDataSource,
-  SystemInventoryProcessesStatesDataSourceRepository,
-} from '../../../../../common/data-source';
-import tableColumns from './table-columns';
-import managedFilters from './managed-filters';
-import { withSystemInventoryProcessesDataSource } from '../../../common/hocs/validate-system-inventory-index-pattern';
-import { ITHygieneInventoryDashboardTable } from '../../../common/components/inventory';
-import { getOverviewProcessesProcessesTab } from './dashboard';
-
-export const ITHygieneProccessesInventoryProcesses =
-  withSystemInventoryProcessesDataSource(props => {
-    return (
-      <ITHygieneInventoryDashboardTable
-        DataSource={SystemInventoryStatesDataSource}
-        DataSourceRepositoryCreator={
-          SystemInventoryProcessesStatesDataSourceRepository
-        }
-        tableDefaultColumns={tableColumns}
-        managedFilters={managedFilters}
-        getDashboardPanels={getOverviewProcessesProcessesTab}
-        tableId='it-hygiene-inventory-processes'
-        indexPattern={props.indexPattern}
-      />
-    );
-  });
